# docs/openapi.yaml
openapi: 3.1.0

info:
  title: Shortener API
  version: 0.21.0
  summary: Create, list and disable short links, and follow them.
  description: |
    Every operation under `/api` needs an access token from the identity provider, and tokens are **bound to the
    client's key** (DPoP, RFC 9449): send `Authorization: DPoP <token>` together with a `DPoP` header holding a proof
    made for that exact request. A plain `Bearer` token is refused, even a valid one. The proof must carry the server's
    **nonce** (`nonce` claim): a request without the current one is answered `401` with `error="use_dpop_nonce"` and the nonce
    in the `DPoP-Nonce` header, and the client repeats it once with a new proof. Every response to a DPoP request carries the
    nonce to use next.

    `deploy/keycloak/DpopClient.java` is a reference implementation of the whole flow, not a library, with only a JDK 17 or
    newer, on your own machine (run `deploy/keycloak/dev-setup` first). For your own program use a DPoP library for its
    language, see docs/OPERATING.md:

    ```
    java deploy/keycloak/DpopClient.java call deploy/keycloak/dev-keys/demo-client.jwk.json demo-client \
      POST http://localhost:8080/api/short-links '{"targetUrl":"https://example.com/some/long/path"}'
    ```

    A client owns the links it creates. A link that belongs to another client answers `404`, as if it did not exist.
    Following a short link (`GET /{shortCode}`) needs no token.

    A link has two addresses. `/api/short-links/{shortCode}` is the resource: read it, change it, the `Location` of a
    created link. `/{shortCode}` is its public face, the URL to share, and is `shortUrl` in the representation.

    Every error is an RFC 9457 problem detail (`application/problem+json`).

    A test checks this file against the controller routes, the configured scope names and the fields of the
    representations, and checks that every operation documents the failures its inputs and its security can cause. It does
    not replay the responses: a status or header described here that no test exercises can still be wrong. Scope names are
    configuration (`shortener.shortlink.scopes.*`); the defaults are shown here.
  license:
    name: Apache-2.0
    identifier: Apache-2.0

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Links
    description: Managing short links. Needs a DPoP-bound access token.
  - name: Redirect
    description: Following a short link. Public.

paths:
  /api/short-links:
    post:
      tags: [Links]
      operationId: createShortLink
      summary: Create a short link
      description: |
        The service picks a random 7-character code. To choose the code, `PUT /api/short-links/{shortCode}`. A `customCode`
        here is a `400`, not ignored. An instance may accept only some target hosts
        (`shortener.shortlink.target-urls.allowed-hosts`), and answers `400` for any other.

        Posting twice creates two links, so a client that lost the answer to a `POST` should look before it repeats it:
        see "Lost answers" in docs/OPERATING.md.
      security:
        - dpop: ['shortlinks:create']
      parameters:
        - $ref: '#/components/parameters/DPoP'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShortLinkRequest'
            examples:
              generated:
                summary: A generated code
                value:
                  targetUrl: https://example.com/some/long/path
      responses:
        '201':
          description: Created. `Location` is the new link; the URL to share is `shortUrl` in the body.
          headers:
            Location:
              description: The link just created, `/api/short-links/{shortCode}`.
              schema:
                type: string
                format: uri
                examples: ['http://localhost:8080/api/short-links/aB3dE9x']
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortLink'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Unavailable'
    get:
      tags: [Links]
      operationId: listShortLinks
      summary: List short links
      description: |
        A client sees its own links, newest first. An administrator sees every client's links, or one client's with
        `createdBy`. The response counts every match (`totalItems`, `totalPages`), so a client can go to any page.
        A page past the last one is empty and still carries the totals. The total is counted separately from the
        page, so a link created between the two can make them differ by one.
      security:
        - dpop: ['shortlinks:read']
        - dpop: ['shortlinks:admin']
      parameters:
        - $ref: '#/components/parameters/DPoP'
        - name: page
          in: query
          description: |
            Zero-based page number. A value that is not a valid page number is not an error: it falls back to the default.
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: |
            Links per page, at most 200. A larger size is cut to 200 and a value that is not a valid size falls back to the
            default, neither is an error. The `size` of the answer is the one that applied.
          schema:
            type: integer
            default: 50
        - name: sort
          in: query
          description: 'Property and direction, `createdAt` or `shortCode`, for example `shortCode,asc`.'
          schema:
            type: string
            default: createdAt,desc
        - name: createdBy
          in: query
          description: Only links created by this client. A client may only name itself.
          schema:
            type: string
      responses:
        '200':
          description: A page of links.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortLinkPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Unavailable'

  /api/short-links/{shortCode}:
    parameters:
      - $ref: '#/components/parameters/ShortCode'
      - $ref: '#/components/parameters/DPoP'
    get:
      tags: [Links]
      operationId: getShortLink
      summary: Read a short link
      description: Includes a disabled link. Another client's link answers `404`.
      security:
        - dpop: ['shortlinks:read']
        - dpop: ['shortlinks:admin']
      responses:
        '200':
          description: The link.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortLink'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Unavailable'
    put:
      tags: [Links]
      operationId: claimShortLink
      summary: Create a short link under a code of your choice
      description: |
        The client chooses the code, so the request names the link and can be repeated. Needs the `claim` scope as well as
        `create`. Codes `api`, `actuator`, `error` and `app` are reserved. An instance may accept only some target hosts
        (`shortener.shortlink.target-urls.allowed-hosts`), and answers `400` for any other.

        - The code is free: the link is created, `201`.
        - The caller already has this link, under this code and for this target: nothing is created, `200` with the link as it
          is now, which may be disabled. This is what a client that lost the answer gets by sending the request again.
        - Anything else on a code that exists is a `409`: another client's link, or the caller's own for another target. A
          link is never overwritten, because its target never changes. A `GET` of the code tells the two apart: the
          caller's own link is read, another client's is a `404`.
      security:
        - dpop: ['shortlinks:create', 'shortlinks:claim']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimShortLinkRequest'
            examples:
              custom:
                summary: A custom code
                value:
                  targetUrl: https://example.com/spring-sale
      responses:
        '200':
          description: The caller already had this link. Nothing was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortLink'
        '201':
          description: Created. `Location` is the new link.
          headers:
            Location:
              description: The link just created, `/api/short-links/{shortCode}`.
              schema:
                type: string
                format: uri
                examples: ['http://localhost:8080/api/short-links/spring-sale']
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortLink'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: The code is reserved, or taken by another client, or by this one for another target.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Unavailable'
    patch:
      tags: [Links]
      operationId: updateShortLink
      summary: Disable a short link
      description: |
        The one change a link takes today is `{"disabled": true}`. The link stops redirecting (`410`) and keeps its code,
        so nobody can register it again. It cannot be enabled again: `{"disabled": false}` is a `400`, as is a body
        without `disabled`. Disabling twice is harmless: the answer is the link as it was first disabled, with the first
        time (`disabledAt`). Other instances can keep redirecting for up to the redirect cache's time to live, 30 seconds
        by default, and an instance that cannot reach the database can keep serving a link it read earlier for up to the
        redirect cache's stale-if-error time, 5 minutes by default.
      security:
        - dpop: ['shortlinks:delete']
        - dpop: ['shortlinks:admin']
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateShortLinkRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateShortLinkRequest'
      responses:
        '200':
          description: The link, disabled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShortLink'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Unavailable'

  /{shortCode}:
    get:
      tags: [Redirect]
      operationId: followShortLink
      summary: Follow a short link
      description: |
        Public. Served from an in-memory cache when the link is active. `HEAD` answers the same status and headers
        without a body. A code that is not 3 to 32 letters, digits, `-` or `_` is a `400`, not a `404`: it could not
        have been issued.
      security: []
      parameters:
        - $ref: '#/components/parameters/ShortCode'
      responses:
        '302':
          description: Redirects to the target URL.
          headers:
            Location:
              description: The target URL.
              schema:
                type: string
                format: uri
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: The link was disabled.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/Unavailable'

components:
  securitySchemes:
    dpop:
      type: oauth2
      description: |
        Client credentials with `private_key_jwt` client authentication, and an access token bound to the client's
        key (DPoP). Present it as `Authorization: DPoP <token>` with a `DPoP` proof header. The token must carry the
        JOSE type `at+jwt`, the audience `shortener-api` and the scopes listed on each operation.
      flows:
        clientCredentials:
          tokenUrl: http://localhost:8180/realms/shortener/protocol/openid-connect/token
          scopes:
            'shortlinks:create': Create links with a generated code.
            'shortlinks:claim': Also choose the code, with `PUT` (needs `shortlinks:create` too).
            'shortlinks:read': Read and list the client's own links.
            'shortlinks:delete': Disable the client's own links.
            'shortlinks:admin': Read and disable any client's links.

  parameters:
    ShortCode:
      name: shortCode
      in: path
      required: true
      description: 3 to 32 letters, digits, `-` or `_`.
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{3,32}$'
        examples: [aB3dE9x]
    DPoP:
      name: DPoP
      in: header
      required: true
      description: |
        A proof JWT (`typ` `dpop+jwt`) made for this request: its method (`htm`), URL without the query (`htu`), the
        hash of the access token (`ath`), a unique `jti`, the time and the server's current nonce (`nonce`). A proof can be
        used once, within 30 seconds.
      schema:
        type: string

  schemas:
    CreateShortLinkRequest:
      type: object
      required: [targetUrl]
      properties:
        targetUrl:
          type: string
          format: uri
          description: An absolute http or https URL.
          examples: ['https://example.com/some/long/path']
        customCode:
          not: {}
          description: Must be absent. A code is chosen with `PUT /api/short-links/{shortCode}`.
    ClaimShortLinkRequest:
      type: object
      required: [targetUrl]
      properties:
        targetUrl:
          type: string
          format: uri
          description: An absolute http or https URL.
          examples: ['https://example.com/spring-sale']
    UpdateShortLinkRequest:
      type: object
      description: A JSON merge patch (RFC 7396). `disabled` is the only field, and it is required for now.
      required: [disabled]
      properties:
        disabled:
          type: boolean
          enum: [true]
          description: Disable the link. `false` is refused, a disabled link cannot be enabled again.
    ShortLink:
      type: object
      required: [shortCode, shortUrl, targetUrl, createdBy, createdAt, disabledAt]
      properties:
        shortCode:
          type: string
          examples: [aB3dE9x]
        shortUrl:
          type: string
          format: uri
          description: |
            The URL to share, `/{shortCode}` on the host the request was made to. It follows the request, so the same link
            read through another name of the instance has another `shortUrl`.
          examples: ['http://localhost:8080/aB3dE9x']
        targetUrl:
          type: string
          format: uri
        createdBy:
          type: [string, 'null']
          description: The client that created the link.
        createdAt:
          type: string
          format: date-time
        disabledAt:
          type: [string, 'null']
          format: date-time
          description: Always present. Null while the link is active.
    ShortLinkPage:
      type: object
      required: [items, page, size, hasNext, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ShortLink'
        page:
          type: integer
        size:
          type: integer
        hasNext:
          type: boolean
          description: Whether another page follows.
        totalItems:
          type: integer
          format: int64
          description: How many links match, over every page.
        totalPages:
          type: integer
          description: How many pages of `size` the matches fill. Zero when nothing matches.
    Problem:
      type: object
      description: An RFC 9457 problem detail.
      required: [title, status]
      properties:
        type:
          type: string
          examples: ['about:blank']
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          description: The path of the request that failed.

  responses:
    BadRequest:
      description: |
        A body that is not valid JSON, lacks a field or has a `customCode` on a `POST`, a short code that is not 3 to 32 letters, digits, `-` or `_`, a sort
        property that is not allowed, a `PATCH` that is not `{"disabled": true}`, or a target whose host this instance does
        not accept.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: |
        No credentials, a bad token or proof, the `Bearer` scheme, or a proof without the current nonce. `WWW-Authenticate`
        names the `DPoP` scheme and the error (`invalid_token`, `invalid_request`, `invalid_dpop_proof` or `use_dpop_nonce`).
        With `use_dpop_nonce`, `DPoP-Nonce` has the nonce for the proof of the repeated request.
      headers:
        WWW-Authenticate:
          schema:
            type: string
            examples: ['DPoP realm="shortener", error="invalid_dpop_proof", algs="ES256 ..."']
        DPoP-Nonce:
          description: The current nonce, on every response to a request under the `DPoP` scheme.
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The token lacks the scope (`WWW-Authenticate` carries `error="insufficient_scope"`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No such link, or it belongs to another client.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TooManyRequests:
      description: Rate limit exceeded (300 a minute per IP, 60 per client).
      headers:
        Retry-After:
          description: Seconds to wait.
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unavailable:
      description: The database cannot be reached; retry after the delay in `Retry-After`.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'